+++ MSIP_Program/src/com/msip/ui/AdminToolsPanel.java
@@ -47,7 +47,7 @@ public class AdminToolsPanel extends JPanel implements ActionListener {
 		StudentPanel studentPanel = new StudentPanel(msipCore, this);
 		AdminPanel adminPanel = new AdminPanel(msipCore, this);
 		reportPanel = new ReportPanel(msipCore, this);
-		NotificationsPanel notificationsPanel = new NotificationsPanel(
+		AdminNotificationTabPanel adminNotificationTabPanel = new AdminNotificationTabPanel(
 				msipCore, this);
 		questionnairePanel = new QuestionnairePanel(msipCore);
 		SettingsPanel settingsPanel = new SettingsPanel(this);
@@ -55,7 +55,7 @@ public class AdminToolsPanel extends JPanel implements ActionListener {
 		tabbedPane.add(studentPanel, GlobalUI.StudentTab);
 		tabbedPane.add(adminPanel, GlobalUI.AdminTab);
 		tabbedPane.add(reportPanel, GlobalUI.ReportTab);
-		tabbedPane.add(notificationsPanel, GlobalUI.NotificationsPanel);
+		tabbedPane.add(adminNotificationTabPanel, GlobalUI.NotificationsPanel);
 		tabbedPane.add(questionnairePanel, GlobalUI.QuestionnairePanel);
 		tabbedPane.add(settingsPanel, GlobalUI.TimePanel);
